Add maxIdx
